requirements.txt
cryptography==41.0.7
bcrypt==3.2.0
Django==3.2.16
django-filter==23.1
pytest==6.2.4
pytest-pythonpath==0.7.3
pytest-django==4.4.0
djangorestframework==3.12.4
djoser
djangorestframework-simplejwt==4.7.2
Pillow==9.3.0
requests==2.26.0
